Refactor vdisk_mgr: move it to separate file.
