+++ tlog/tlogserver/server/server.go
@@ -25,11 +25,11 @@ type Server struct {
 	listener             net.Listener
 	flusherConf          *flusherConfig
 	vdiskMgr             *vdiskManager
-	globalConf           *config.Config
+	fileConf             *config.Config
 }
 
 // NewServer creates a new tlog server
-func NewServer(globalConf *config.Config, conf *Config, poolFactory tlog.RedisPoolFactory) (*Server, error) {
+func NewServer(conf *Config, poolFactory tlog.RedisPoolFactory) (*Server, error) {
 	if conf == nil {
 		return nil, errors.New("tlogserver requires a non-nil config")
 	}
@@ -73,7 +73,7 @@ func NewServer(globalConf *config.Config, conf *Config, poolFactory tlog.RedisPo
 		flusherConf:          flusherConf,
 		maxRespSegmentBufLen: schema.RawTlogRespLen(conf.FlushSize),
 		vdiskMgr:             newVdiskManager(conf.AggMq, conf.BlockSize, conf.FlushSize, conf.ConfigPath),
-		globalConf:           globalConf,
+		fileConf:             conf.FileConfig,
 	}, nil
 }
 
@@ -160,7 +160,7 @@ func (s *Server) handshake(r io.Reader, w io.Writer, conn *net.TCPConn) (vd *vdi
 		return // error return
 	}
 
-	vd, err = s.vdiskMgr.Get(s.globalConf, vdiskID, req.FirstSequence(), s.createFlusher, conn, s.flusherConf)
+	vd, err = s.vdiskMgr.Get(s.fileConf, vdiskID, req.FirstSequence(), s.createFlusher, conn, s.flusherConf)
 	if err != nil {
 		status = tlog.HandshakeStatusInternalServerError
 		err = fmt.Errorf("couldn't create vdisk %s: %s", vdiskID, err.Error())
